% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap.R
\name{is_leap}
\alias{is_leap}
\title{Check if a Year is a Leap Year}
\usage{
is_leap(year)
}
\arguments{
\item{year}{An integer representing a year.}
}
\value{
A logical value: \code{TRUE} if the year is a leap year, \code{FALSE} otherwise.
}
\description{
This function determines whether a given year is a leap year according to the Gregorian calendar rules:
\itemize{
\item Divisible by 4: Yes
\item Divisible by 100: No, unless also divisible by 400
}
}
\examples{
is_leap(1992)  # TRUE
is_leap(1900)  # FALSE
is_leap(2000)  # TRUE
is_leap(2023)  # FALSE

}
